refactor: components optimization
