+++ packages/components/checkbox/consts/index.ts
