+++ packages/components/collapse/collapse.tsx
@@ -8,7 +8,6 @@ import { usePrefixClass } from '../hooks/useConfig';
 export default defineComponent({
   name: 'TCollapse',
   props,
-
   setup(props: TdCollapseProps) {
     const componentName = usePrefixClass('collapse');
     const borderlessClass = usePrefixClass('-border-less');
